requirements.txt
# Core Python
python-dotenv==1.0.1
numpy==1.26.4
pandas==2.2.3
scikit-learn==1.5.2
scipy==1.14.1
joblib==1.4.2

# Machine Learning Models
xgboost==2.1.1
catboost==1.2.5
shap==0.46.0

# Visualization (EDA + Dashboard)
matplotlib==3.9.2
seaborn==0.13.2
plotly==5.24.1

# Backend (API)
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.0
pydantic-settings==2.4.0

# Frontend (Streamlit Dashboard)
streamlit==1.38.0
streamlit-shap==1.0.2

# Database
SQLAlchemy==2.0.35
psycopg2-binary==2.9.9
alembic==1.13.3

# Utility
tqdm==4.66.5
requests==2.32.3
uuid==1.30
typing-extensions==4.12.2

# Testing & Quality
pytest==8.3.3
pytest-cov==5.0.0
flake8==7.1.1
black==24.10.0
isort==5.13.2
mypy==1.12.0
pre-commit==3.8.0

# Serialization
pickle-mixin==1.0.2
cloudpickle==3.1.0

# Deployment / Docker
gunicorn==23.0.0
